# ♻️ refactor: rustify writer utility `test_free_rows`