--- a/packages/biliass/rust/src/lib.rs
+++ b/packages/biliass/rust/src/lib.rs
@@ -31,5 +31,7 @@ fn biliass_pyo3(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(python::py_convert_color, m)?)?;
     m.add_function(wrap_pyfunction!(python::py_get_zoom_factor, m)?)?;
     m.add_function(wrap_pyfunction!(python::py_convert_flash_rotation, m)?)?;
+    m.add_function(wrap_pyfunction!(python::py_test_free_rows, m)?)?;
+    m.add_class::<python::PyRows>()?;
     Ok(())
 }