--- a/packages/biliass/rust/src/python/comment.rs
+++ b/packages/biliass/rust/src/python/comment.rs
@@ -27,7 +27,7 @@ impl PyCommentPosition {
 
 #[pyclass(name = "Comment", frozen)]
 pub struct PyComment {
-    inner: comment::Comment,
+    pub inner: comment::Comment,
 }
 
 impl PyComment {
@@ -96,7 +96,7 @@ impl PyComment {
 
 #[pyclass(name = "OptionComment", frozen)]
 pub struct PyOptionComment {
-    inner: Option<comment::Comment>,
+    pub inner: Option<comment::Comment>,
 }
 
 impl PyOptionComment {
@@ -107,27 +107,27 @@ impl PyOptionComment {
 
 #[pymethods]
 impl PyOptionComment {
-    fn is_none(&self) -> bool {
+    pub fn is_none(&self) -> bool {
         self.inner.is_none()
     }
 
-    fn is_some(&self) -> bool {
+    pub fn is_some(&self) -> bool {
         self.inner.is_some()
     }
 
-    fn unwrap(&self) -> PyComment {
+    pub fn unwrap(&self) -> PyComment {
         PyComment::new(self.inner.clone().unwrap())
     }
 
     #[staticmethod]
-    fn from_comment(comment: &PyComment) -> Self {
+    pub fn from_comment(comment: &PyComment) -> Self {
         PyOptionComment {
             inner: Some(comment.inner.clone()),
         }
     }
 
     #[staticmethod]
-    fn none() -> Self {
+    pub fn none() -> Self {
         PyOptionComment { inner: None }
     }