--- a/packages/biliass/rust/src/python/mod.rs
+++ b/packages/biliass/rust/src/python/mod.rs
@@ -8,5 +8,5 @@ pub use proto::{PyDanmakuElem, PyDmSegMobileReply};
 pub use reader::{py_read_comments_from_protobuf, py_read_comments_from_xml};
 pub use writer::{
     py_ass_escape, py_convert_color, py_convert_flash_rotation, py_convert_timestamp,
-    py_get_zoom_factor,
+    py_get_zoom_factor, py_test_free_rows, PyRows,
 };