--- a/packages/biliass/rust/src/writer/mod.rs
+++ b/packages/biliass/rust/src/writer/mod.rs
@@ -1,1 +1,2 @@
+pub mod rows;
 pub mod utils;